Moved card related functions to the enum
