+++ src/models/card.rs
@@ -16,23 +16,23 @@ pub enum Card {
 }
 
 macro_rules! impl_various {
-    ($($var:ident),*) => {
+    ($($card_variant:ident),*) => {
         impl Card {
             pub fn validate_answer(&mut self) -> UserAnswer {
                 match self {
-                    $(Card::$var(card) => card.validate_answer(),)*
+                    $(Card::$card_variant(card) => card.validate_answer()),*
                 }
             }
 
             pub fn check_answered(&mut self) -> bool {
                 match self {
-                    $(Card::$var(card) => card.user_answer != UserAnswer::Undecided,)*
+                    $(Card::$card_variant(card) => card.user_answer != UserAnswer::Undecided),*
                 }
             }
 
             pub fn instructions(&self) -> String {
                 match self {
-                    $(Card::$var(card) => card.instructions(),)*
+                    $(Card::$card_variant(card) => card.instructions()),*
                 }
             }
         }
@@ -40,7 +40,7 @@ macro_rules! impl_various {
         impl fmt::Display for Card {
             fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
                 match self {
-                    $(Card::$var(card) => write!(f, "{card}"),)*
+                    $(Card::$card_variant(card) => write!(f, "{card}")),*
                 }
             }
         }
@@ -54,3 +54,50 @@ impl_various!(
     FillInTheBlanks,
     Order
 );
+
+
+#[derive(Debug)]
+pub enum ParsingError {
+    NoCardType,
+}
+
+impl Card {
+    pub fn extract_card_title(content: &String) -> (String, String) {
+        // Don't unwrap
+        let question = content.lines().nth(0).unwrap()[1..].trim().to_string();
+        let content = content.lines().skip(1).collect::<Vec<&str>>().join("\n");
+
+        (question, content)
+    }
+
+    pub fn card_parser(content: String) -> Result<Vec<Self>, ParsingError> {
+        let cards: Vec<Card> = content
+            .split("---")
+            .map(|section| {
+                let sections = section
+                    .trim()
+                    .split("\n\n")
+                    .filter(|item| !item.is_empty())
+                    .collect::<Vec<&str>>();
+
+                match sections[0].to_lowercase().as_str() {
+                    "flashcard" => Card::FlashCard(FlashCard::parse_raw(sections[1].to_string())),
+                    "multiple_choice" => {
+                        Card::MultipleChoice(MultipleChoice::parse_raw(sections[1].to_string()))
+                    }
+                    "multiple_answer" => {
+                        Card::MultipleAnswer(MultipleAnswer::parse_raw(sections[1].to_string()))
+                    }
+                    "fill_in_the_blanks" => {
+                        Card::FillInTheBlanks(FillInTheBlanks::parse_raw(sections[1].to_string()))
+                    }
+                    "order" => Card::Order(Order::parse_raw(sections[1].to_string())),
+                    // Replace with ParsingError datatype
+                    _ => panic!("Parsing Error"),
+                }
+            })
+            .collect();
+
+        Ok(cards)
+    }
+}
