+++ src/models/card_types/multiple_answer.rs
@@ -1,7 +1,7 @@
 use core::fmt;
 
 use crate::{
-    extract_card_title,
+    Card,
     models::{choice::Choice, stateful_list::StatefulList},
     UserAnswer,
 };
@@ -39,7 +39,7 @@ impl MultipleAnswer {
     }
 
     pub fn parse_raw(content: String) -> Self {
-        let (question, content) = extract_card_title(&content);
+        let (question, content) = Card::extract_card_title(&content);
 
         Self {
             question,
